backend/requirements.txt
flask>=2.0
flask-cors
facenet-pytorch>=2.6.0
torch              
torchvision
opencv-python
numpy==1.26.4
Pillow
